+++ ChangeLog.md
@@ -3,3 +3,8 @@ Commit on 19-02-21. Author - Karen
 * RESTful web service 
 * Repository using JPARepository
 * properties for hibernate metrics and queries added
+
+Commit on 23-02-21. Author - Karen
+----------------------------------
+* Entity and POJO separated
+* Object mapper implemented